requirements.txt
python-telegram-bot==20.7
pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
urllib3>=1.26.0
google-generativeai>=0.3.0
httpx>=0.24.0